% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_AFINN}
\alias{data_dictionary_AFINN}
\title{Nielsen's (2011) 'new ANEW' valenced word list}
\format{
A \link{dictionary} with one key, \code{AFINN}, with valences from -5 (negative) to +5
(positive).
}
\source{
\url{http://www2.imm.dtu.dk/pubdb/views/publication_details.php?id=6010}
}
\usage{
data_dictionary_AFINN
}
\description{
A \pkg{quanteda} \link[quanteda:dictionary]{dictionary} object containing Finn Årup
Nielsen's (2011) 'new ANEW' valenced word list, a publicly available list of
English words rated for valence with values between -5 (negative) and +5
(positive). AFINN-111, the latest version, contains 2,477 words and phrases.
}
\section{License}{

\href{http://www.opendatacommons.org/licenses/odbl/1.0/}{Open Database License (ODbL) v1.0}
}

\references{
Nielsen, F. Å. (2011). \href{https://arxiv.org/abs/1103.2903}{A new ANEW: Evaluation of a Word List for Sentiment Analysis in Microblogs.} In
\emph{Proceedings of the ESWC2011 Workshop on 'Making Sense of Microposts':
Big Things Come in Small Packages}, 93--98.
}
\keyword{data}
