% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_ANSW}
\alias{data_dictionary_ANSW}
\title{Affective Norms for Spanish Words (ANSW)}
\format{
A \link[quanteda:dictionary]{quanteda::dictionary} with eight valenced keys: \code{valence}, \code{arousal}, \code{concreteness},
\code{happiness}, \code{anger}, \code{sadness}, \code{fearness}, and \code{disgust}, each containing
the same 875 Spanish word values with associated valences.
}
\source{
\url{https://link.springer.com/article/10.3758/s13428-015-0572-5}
}
\usage{
data_dictionary_ANSW
}
\description{
A \pkg{quanteda} \link[quanteda:dictionary]{dictionary} object containing the
ANSW, or Affective Norms for Spanish Words (Hinojosa et al. 2016) valenced
lexicon. The ANSW provides a lexicon of 875 distinct Spanish words that are
associated with eight valenced categories: valence, arousal, concreteness,
and five discrete emotional categories (happiness, anger, sadness, fear,
and disgust).
}
\note{
This dictionary was created by Mariano Rico (mariano.rico@upm.es). For
questions about the dictionary data, please consult the original publication.
}
\section{License}{

Unknown.
}

\references{
Hinojosa, J.A., Martínez-García, N., Villalba-García, C., Fernández-Folgueiras,
U., & Sánchez-Carmona, A. (2016). Affective norms of 875 Spanish words for
five discrete emotional categories and two emotional dimensions.
\emph{Behavior Research Methods}, 48, 272--284.
\doi{10.3758/s13428-015-0572-5}
}
\keyword{data}
