% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_ANSW2018}
\alias{data_dictionary_ANSW2018}
\title{Affective Norms for Spanish Words (ANSW) 2018}
\format{
A \link[quanteda:dictionary]{quanteda::dictionary} with seven valenced keys: \code{valence}, \code{arousal},
\code{happiness}, \code{anger}, \code{sadness}, \code{fearness}, and \code{disgust}, each containing
the same 10,491 Spanish word values with associated valences.
}
\source{
\url{https://link.springer.com/article/10.3758/s13428-017-0962-y}
}
\usage{
data_dictionary_ANSW2018
}
\description{
A \pkg{quanteda} \link[quanteda:dictionary]{dictionary} object containing the
ANSW 2018, or Affective Norms for Spanish Words (Stadthagen-González H. et al.) valenced
lexicon. The ANSW 2018 provides a lexicon of 10,491 distinct Spanish words (x10 2016 version) that are
associated with seven (the 2016 version was for eight) valenced categories: valence, arousal,
and five discrete emotional categories (happiness, anger, sadness, fear,
and disgust).
}
\note{
This dictionary was created by Mariano Rico (mariano.rico@upm.es). For
questions about the dictionary data, please consult the original publication.
}
\section{License}{

For non-profit academic research purposes.
}

\references{
Stadthagen-González, H., Ferré, P., Pérez-Sánchez, M.A. et al. Norms
for 10,491 Spanish words for five discrete emotions: Happiness, disgust,
anger, fear, and sadness.
\emph{Behavior Research Methods} 50, 1943–1952.
\doi{10.3758/s13428-017-0962-y}
}
\keyword{data}
