% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_Chen2014}
\alias{data_dictionary_Chen2014}
\title{Spanish polar dictionary (Chen2014)}
\format{
A \link[quanteda:dictionary]{quanteda::dictionary} with two keys: \code{pos}, and \code{neg}.
}
\source{
\url{https://www.kaggle.com/rtatman/sentiment-lexicons-for-81-languages}
}
\usage{
data_dictionary_Chen2014
}
\description{
A \pkg{quanteda} \link[quanteda:dictionary]{dictionary} object containing
positive and negative words for the Spanish language. The dictionary provides
1,555 positive words and 2,720 negative words. The dataset can be found in
Kaggle.
}
\note{
This dictionary was created by Mariano Rico (mariano.rico@upm.es). For
questions about the dictionary data, please consult the original publication.
}
\section{License}{

\href{https://www.gnu.org/licenses/gpl-3.0.html}{GPL-3}.
}

\references{
Chen, Yanqing and Steven Skiena. (2014) "Building sentiment lexicons for all
major languages." In \emph{Proceedings of the 52nd annual meeting of the
Association for Computational Linguistics} (volume 2: short papers), pp.
383-389.
}
\keyword{data}
