% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_ChenSkiena2014}
\alias{data_dictionary_ChenSkiena2014}
\title{Chen and Skiena's (2014) sentiment dictionaries for 81 languages}
\format{
For a given language xx, a \link[quanteda:dictionary]{dictionary} with two keys: \code{xx_pos}, and \code{xx_neg}, where \code{xx} is the two-character language code, corresponding to the ISO 3166-1 A-2 standard.
}
\source{
\url{https://www.kaggle.com/rtatman/sentiment-lexicons-for-81-languages}
}
\usage{
data_dictionary_ChenSkiena2014
}
\description{
A \pkg{quanteda} \link[quanteda:dictionary]{dictionary} object containing
positive and negative words for 81 (non-English) languages. The dataset can
be found in Kaggle (as well as referenced from the original publication
page).
}
\details{
The languages are:\tabular{ll}{
   \strong{Code} \tab \strong{Language} \cr
   af \tab Afrikaans \cr
   an \tab Aragonese \cr
   ar \tab Arabic \cr
   az \tab Azerbaijani \cr
   be \tab Belarusian \cr
   bg \tab Bulgarian \cr
   bn \tab Bengali \cr
   br \tab Breton \cr
   bs \tab Bosnian \cr
   ca \tab Catalan; Valencian \cr
   cs \tab Czech \cr
   cy \tab Welsh \cr
   da \tab Danish \cr
   de \tab German \cr
   el \tab Greek, Modern \cr
   eo \tab Esperanto \cr
   es \tab Spanish; Castilian \cr
   et \tab Estonian \cr
   eu \tab Basque \cr
   fa \tab Persian \cr
   fi \tab Finnish \cr
   fo \tab Faroese \cr
   fr \tab French \cr
   fy \tab Western Frisian \cr
   ga \tab Irish \cr
   gd \tab Scottish Gaelic; Gaelic \cr
   gl \tab Galician \cr
   gu \tab Gujarati \cr
   he \tab Hebrew (modern) \cr
   hi \tab Hindi \cr
   hr \tab Croatian \cr
   ht \tab Haitian; Haitian Creole \cr
   hu \tab Hungarian \cr
   hy \tab Armenian \cr
   ia \tab Interlingua \cr
   id \tab Indonesian \cr
   io \tab Ido \cr
   is \tab Icelandic \cr
   it \tab Italian \cr
   ka \tab Georgian \cr
   km \tab Khmer \cr
   kn \tab Kannada \cr
   ko \tab Korean \cr
   ku \tab Kurdish \cr
   ky \tab Kirghiz, Kyrgyz \cr
   la \tab Latin \cr
   lb \tab Luxembourgish, Letzeburgesch \cr
   lt \tab Lithuanian \cr
   lv \tab Latvian \cr
   mk \tab Macedonian \cr
   mr \tab Marathi (Marāṭhī) \cr
   ms \tab Malay \cr
   mt \tab Maltese \cr
   nl \tab Dutch \cr
   nn \tab Norwegian Nynorsk \cr
   no \tab Norwegian \cr
   pl \tab Polish \cr
   pt \tab Portuguese \cr
   rm \tab Romansh \cr
   ro \tab Romanian, Moldavian, Moldovan \cr
   ru \tab Russian \cr
   sk \tab Slovak \cr
   sl \tab Slovene \cr
   sq \tab Albanian \cr
   sr \tab Serbian \cr
   sv \tab Swedish \cr
   sw \tab Swahili \cr
   ta \tab Tamil \cr
   te \tab Telugu \cr
   th \tab Thai \cr
   tk \tab Turkmen \cr
   tl \tab Tagalog \cr
   tr \tab Turkish \cr
   uk \tab Ukrainian \cr
   ur \tab Urdu \cr
   uz \tab Uzbek \cr
   vi \tab Vietnamese \cr
   vo \tab Volapük \cr
   wa \tab Walloon \cr
   yi \tab Yiddish \cr
   zh \tab Chinese \cr
   zhw \tab Chinese (characters and words) \cr
}
}
\note{
This dictionary was created by Mariano Rico (mariano.rico@upm.es). For
questions about the dictionary data, please consult the original publication.
}
\section{License}{

\href{https://www.gnu.org/licenses/gpl-3.0.html}{GPL-3}.
}

\examples{
library("quanteda.sentiment")
# Tests for 4 Spain languages

# Test for Spanish
polarity(data_dictionary_ChenSkiena2014) <- list(neg = "es_neg", pos = "es_pos")
textstat_polarity("Hay que ser cruel para ser amable.", 
                  dictionary = data_dictionary_ChenSkiena2014) 
##   doc_id sentiment
## 1  text1 0.5108256

# Test for Catalan
polarity(data_dictionary_ChenSkiena2014) <- list(neg = "ca_neg", pos = "ca_pos")
textstat_polarity("Cal ser cruel per ser amable.", 
                  dictionary = data_dictionary_ChenSkiena2014)
##   doc_id sentiment
## 1  text1 0.5108256

# Test for Basque
polarity(data_dictionary_ChenSkiena2014) <- list(neg = "eu_neg", pos = "eu_pos")
textstat_polarity("Krudel izan behar da atsegina izateko.", 
                  dictionary = data_dictionary_ChenSkiena2014)
##   doc_id sentiment
## 1  text1  1.098612

# Test for Galician
polarity(data_dictionary_ChenSkiena2014) <- list(neg = "gl_neg", pos = "gl_pos")
textstat_polarity("Hai que ser cruel para ser amable.", 
                  dictionary = data_dictionary_ChenSkiena2014)
##   doc_id sentiment
## 1  text1 0.5108256
}
\references{
Chen, Yanqing and Steven Skiena. (2014) "Building sentiment lexicons for all
major languages." In \emph{Proceedings of the 52nd annual meeting of the
Association for Computational Linguistics} (volume 2: short papers), pp.
383-389. \doi{10.3115/v1/P14-2063}
}
\keyword{data}
