% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_geninqposneg}
\alias{data_dictionary_geninqposneg}
\title{Augmented General Inquirer \emph{Positiv} and \emph{Negativ} dictionary}
\format{
A \link{dictionary} of fixed word patterns with two keys:
\itemize{
\item \code{positive}: 1,653 words with positive polarity
\item \code{negative}: 2,010 words with negative polarity
}
}
\source{
\url{http://www.wjh.harvard.edu/~inquirer/spreadsheet_guide.htm}
}
\usage{
data_dictionary_geninqposneg
}
\description{
A \pkg{quanteda} \link[quanteda:dictionary]{dictionary} object containing the
\emph{Positiv} and \emph{Negativ} dictionary entries from the augmented
General Inquirer. These are new valence categories described at
\url{http://www.wjh.harvard.edu/~inquirer/homecat.htm} but also including the
terms from the \href{http://www.wjh.harvard.edu/~inquirer/Yes.html}{yes} and
\href{http://www.wjh.harvard.edu/~inquirer/No.html}{no} dictionary entries.
}
\references{
Stone, P.J., Dunphy, C.D., & Smith, M.S. (1966).
\emph{The General Inquirer: A Computer Approach to Content Analysis.}
Cambridge, MA: MIT Press.
}
\keyword{data}
