% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda.sentiment-package.R
\docType{package}
\name{quanteda.sentiment-package}
\alias{quanteda.sentiment}
\alias{quanteda.sentiment-package}
\title{quanteda.sentiment: Sentiment analysis using quanteda}
\description{
Adds functions and dictionaries for computing sentiment using the 'quanteda' package.
}
\author{
\strong{Maintainer}: Kenneth Benoit \email{kbenoit@lse.ac.uk} [copyright holder]

}
\keyword{internal}
