% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-documentation.R
\docType{data}
\name{data_dictionary_ANEW}
\alias{data_dictionary_ANEW}
\title{Affective Norms for English Words}
\format{
An object of class \code{dictionary3} of length 3.
}
\usage{
data_dictionary_ANEW
}
\description{
TBC
}
\keyword{data}
