% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dictionary3-class.R
\docType{class}
\name{dictionary3-class}
\alias{dictionary3-class}
\alias{print,dictionary3-method}
\alias{print.dictionary3}
\alias{[,dictionary3,index,ANY,ANY-method}
\alias{[[,dictionary3,index-method}
\title{dictionary3 class}
\usage{
\S4method{print}{dictionary3}(
  x,
  max_nkey = quanteda_options("print_dictionary_max_nkey"),
  max_nval = quanteda_options("print_dictionary_max_nval"),
  show_summary = quanteda_options("print_dictionary_summary"),
  ...
)

\S4method{[}{dictionary3,index,ANY,ANY}(x, i)

\S4method{[[}{dictionary3,index}(x, i)
}
\arguments{
\item{max_nkey}{max number of keys to print; default is from the
\code{print_dictionary_max_max_nkey} setting of \code{\link[=quanteda_options]{quanteda_options()}}}

\item{max_nval}{max number of values to print; default is from the
\code{print_dictionary_max_nval} setting of \code{\link[=quanteda_options]{quanteda_options()}}}

\item{i}{index for entries}

\item{object}{the dictionary to be extracted}
}
\description{
This extends the \pkg{quanteda} \code{dictionary2} class with a version that
recognizes \link{valence} and \link{polarity} meta-data for sentiment analysis.
}
\section{Slots}{

\describe{
\item{\code{.Data}}{named list of mode character, where each element name is a
dictionary "key" and each element is one or more dictionary entry "values"
consisting of a pattern match}

\item{\code{meta}}{list of object metadata}
}}

\keyword{dictionary}
\keyword{internal}
