% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quanteda.sentiment-package.R
\docType{package}
\name{quanteda.sentiment-package}
\alias{quanteda.sentiment}
\alias{quanteda.sentiment-package}
\title{quanteda.sentiment: Sentiment Analysis using 'quanteda'}
\description{
Adds functions and dictionaries for computing sentiment using the 'quanteda' package.
}
\author{
\strong{Maintainer}: Kenneth Benoit \email{kbenoit@smu.edu.sg} (\href{https://orcid.org/0000-0002-0797-564X}{ORCID})

Other contributors:
\itemize{
  \item Mariano Rico \email{mariano.rico@upm.es} (\href{https://orcid.org/0000-0001-5878-8521}{ORCID}) (Contributor of Spanish emotion dictionary) [contributor]
}

}
\keyword{internal}
